--- a/Sources/SwordGenerator/Parser/SwordParser.swift
+++ b/Sources/SwordGenerator/Parser/SwordParser.swift
@@ -19,43 +19,51 @@ package struct SwordParser {
     let moduleRegistry = ModuleRegistry()
     let importRegistry = ImportRegistry()
 
-    for sourceFile in sourceFiles {
-      let visitors: [SyntaxVisitor] = [
-        ComponentVisitor(
-          componentRegistry: componentRegistry,
-          sourceFile: sourceFile
-        ),
-        SubcomponentVisitor(
-          componentRegistry: componentRegistry,
-          sourceFile: sourceFile
-        ),
-        DependencyVisitor(
-          dependencyRegistry: dependencyRegistry,
-          sourceFile: sourceFile
-        ),
-        ModuleVisitor(
-          moduleRegistry: moduleRegistry,
-          sourceFile: sourceFile
-        ),
-        ImportVisitor(
-          importRegistry: importRegistry,
-          sourceFile: sourceFile
-        ),
-      ]
-      for visitor in visitors {
-        visitor.walk(sourceFile.tree)
+    Logging.recordInterval(name: "parseSourceFiles") {
+      for sourceFile in sourceFiles {
+        let visitors: [SyntaxVisitor] = [
+          ComponentVisitor(
+            componentRegistry: componentRegistry,
+            sourceFile: sourceFile
+          ),
+          SubcomponentVisitor(
+            componentRegistry: componentRegistry,
+            sourceFile: sourceFile
+          ),
+          DependencyVisitor(
+            dependencyRegistry: dependencyRegistry,
+            sourceFile: sourceFile
+          ),
+          ModuleVisitor(
+            moduleRegistry: moduleRegistry,
+            sourceFile: sourceFile
+          ),
+          ImportVisitor(
+            importRegistry: importRegistry,
+            sourceFile: sourceFile
+          ),
+        ]
+        for visitor in visitors {
+          visitor.walk(sourceFile.tree)
+        }
+      }
+
+      for target in targets {
+        importRegistry.register(target)
       }
     }
 
-    for target in targets {
-      importRegistry.register(target)
+    let componentValidationResult = Logging.recordInterval(name: "makeComponentTree") {
+      ComponentValidator(componentRegistry: componentRegistry)
+        .validate()
     }
-
-    let componentValidationResult = ComponentValidator(componentRegistry: componentRegistry)
-      .validate()
-    let dependencyValidationResult = DependencyValidator(dependencyRegistry: dependencyRegistry)
-      .validate()
-    let moduleValidationResult = ModuleValidator(moduleRegistry: moduleRegistry).validate()
+    let dependencyValidationResult = Logging.recordInterval(name: "makeDependencies") {
+      DependencyValidator(dependencyRegistry: dependencyRegistry)
+        .validate()
+    }
+    let moduleValidationResult = Logging.recordInterval(name: "makeModules") {
+      ModuleValidator(moduleRegistry: moduleRegistry).validate()
+    }
     guard
       case .valid((let component, let subcomponentsByParent)) = componentValidationResult,
       case .valid(let dependenciesByComponentName) = dependencyValidationResult,
@@ -72,19 +80,23 @@ package struct SwordParser {
       exit(1)
     }
 
-    let bindingGraphFactory = BindingGraphFactory(
-      subcomponentsByParent: subcomponentsByParent,
-      dependenciesByComponentName: dependenciesByComponentName,
-      modulesByComponentName: modulesByComponentName
-    )
-    let bindingGraph = bindingGraphFactory.makeBindingGraph(rootComponent: component)
-    let bindingGraphValidationResult = BindingGraphValidator(bindingGraph: bindingGraph).validate()
-
-    guard case .valid = bindingGraphValidationResult else {
-      for report in bindingGraphValidationResult.reports {
-        reporter.send(report)
+    let bindingGraph = Logging.recordInterval(name: "makeBindingGraph") {
+      let bindingGraphFactory = BindingGraphFactory(
+        subcomponentsByParent: subcomponentsByParent,
+        dependenciesByComponentName: dependenciesByComponentName,
+        modulesByComponentName: modulesByComponentName
+      )
+      let bindingGraph = bindingGraphFactory.makeBindingGraph(rootComponent: component)
+      let bindingGraphValidationResult = BindingGraphValidator(bindingGraph: bindingGraph)
+        .validate()
+      guard case .valid = bindingGraphValidationResult else {
+        for report in bindingGraphValidationResult.reports {
+          reporter.send(report)
+        }
+        exit(1)
       }
-      exit(1)
+
+      return bindingGraph
     }
 
     return (